Fixed Camera Controls and more
Made camera controls functional. Press the arrow keys to move the camera in that direction. It will always keep the player in the center of the screen.
Added a LookAtTarget script that makes the camera look at the player.
Made most of the non-interactable objects in the skatepark have the "ground" tag so the play can jump again after touching them. Ramps, Pipes, and Rails mostly. Added another pickup over the button in the first area.