Implemented Narration VO and Pickup VO
Implemented functioning Narration at the start of each level, and when the player picks up a pickup. Only issue is that the pickup VO won't play on the Skatepark level for some reason. Also the door moving script doesn't work so I've kicked in the door so the player can advance through the level.